Vehicle Routes & Paladin Image Links
Implemented scraper for 503 Bus and Campus Shuttle routes and stops, including the capacity to determine where on the route a stop fell.

Implemented calculations for how far from each stop on its route a vehicle currently was.

Fixed problem with 503 Bus' ID changing every couple of days by scraping it from the web.

Implemented method to scrape article images from the Paladin, given a link to the article.