Local Database, sqlite adjustments
I created a local sqlite database to use while debugging the hours scraper bug. Currently it only has the buildingHours and buildingLocations tables. Trying to determine now why table is so large, hours scraper doesn't add more to it when executed, it may be a different scraper. Also some names in buildingLocations are outdated.