improved routes, wrapper class, and slowly updating database to match old one
This will take a while to get all the old tables and triggers set up in the new database. I have access to everything I need to do this so it shouldn't be too bad. I'm going to set up the triggers and see if I can't figure out how to automate updates after all of this.