The database has been tweaked to work with all scrapers. All scrapers have been run and most work.
Changed file paths and a few other minor code changes have been made for the scrapers to work. Currently there are a few bugs in the scrapers. First, the greenlink websites have moved so we will want to find a new one for that. Also, I need to figure out how to better the sql qeuries on the vehicle scraper, they don't execute yet.